requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
asgiref==3.4.1
certifi==2021.5.30
dj-database-url==0.5.0
Django==3.2.5
django-heroku==0.3.1
gunicorn==20.1.0
numpy==1.21.1
Pillow==8.3.1
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.1
torch==1.8.1+cpu
torchvision==0.9.1+cpu
typing-extensions==3.10.0.0
whitenoise==5.3.0
wincertstore==0.2
